Updated quote nesting
